requirements.txt
absl-py==2.2.1
babel==2.17.0
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
immutabledict==4.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.4
ortools==9.12.4544
pandas==2.2.3
protobuf==5.29.4
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.39
tkcalendar==1.6.1
typing_extensions==4.13.0
tzdata==2025.2
Werkzeug==3.1.3
